8.7.0
Introduced a new `IsEmpty` property to the `ISnowflake` interface
to determine if an identifier is empty (uninitialized or default).
Implemented the `IsEmpty` property in the `Snowflake` class,
which checks if the internal `__combined` field equals zero.
This enhances the functionality of the `ISnowflake` interface
and its concrete implementation in the `Snowflake` class.