5.4.0
Updated the singleton initialization of the NLogix logger in `NLogix.Host.cs` to use a static local function for improved clarity and performance. The configuration of default logging outputs has been structured within the lambda function.

Enhanced the documentation in `NLogix.cs` by adding a new paragraph that clarifies the purpose of the NLogix logging engine, emphasizing its high-performance, extensible nature and its implementation of the ILogger interface for unified logging.