6.7.0
Removed the `CipherSuiteType` parameter from the `Decrypt` method in the `IPacketEncryptor` interface and related delegates to streamline the decryption process. Updated the `Decrypt` delegate in `PacketTransformer` and `decryptDel` in `PacketCatalogFactory` to reflect this change. Refactored all usages of the `Decrypt` method, including in `UnwrapPacketMiddleware`. Improved XML documentation formatting for clarity and consistency.