6.7.5
Updated the `Encrypt` method in `Salsa20.cs` to include the creation of a `ciphertext` array and invocation of the `B1C2D3E4` method, returning the `ciphertext`. Reformatted the `Decrypt` method for readability without altering functionality.

Removed the copyright notice and unused `using Nalix.Framework.Cryptography.Primitives;` directive in `Speck.cs` for cleanup.